pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-email-assistant"
version = "0.1.0"
description = "AI-powered email assistant for OpenClaw with meeting scheduling and bounce handling"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Yuchuan Xu", email = "yuchuanxu@example.com" }]
keywords = ["openclaw", "email", "automation", "ai", "gmail", "calendar"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Home Automation",
]
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "pyyaml>=6.0",
    "Jinja2>=3.0.0",
    "notion-client>=2.0.0",
    "openclaw>=0.4.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/openclaw-email-assistant"
Repository = "https://github.com/yourusername/openclaw-email-assistant"
Issues = "https://github.com/yourusername/openclaw-email-assistant/issues"

[project.scripts]
# No CLI entry point; this is an OpenClaw skill

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.jinja2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.flake8]
max-line-length = 127
exclude = [".git", "__pycache__", "venv", "dist", "build"]
